# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/users/me": {
      "get": {
        "operationId": "users_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["users"]
      },
      "patch": {
        "operationId": "users_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["users"]
      }
    },
    "/api/v1/users/me/profile": {
      "get": {
        "operationId": "users_myProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Identity + roles + per-role counters for the profile screen, in one request.",
        "tags": ["users"]
      }
    },
    "/api/v1/users/me/roles/scout": {
      "post": {
        "operationId": "users_becomeScout",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Become a scout — the one role a user may give themselves.\n\nEverything else is granted by somebody: `player` comes with creating a\nprofile, `coach` from the academy that hires you, `academy_manager` and\n`admin` from an admin. Scouting needs no permission because it starts with\nzero authority: a new scout's recommendations carry the lowest §1.5 weight\nand only earn more by being accepted.",
        "tags": ["users"]
      }
    },
    "/api/v1/users/me/avatar/upload-url": {
      "post": {
        "operationId": "users_avatarUploadUrl",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AvatarUploadUrlDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["users"]
      }
    },
    "/api/v1/users/me/contact/request": {
      "post": {
        "operationId": "users_requestContactChange",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestContactChangeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Step 1 of changing a phone or email: prove you control the new destination.",
        "tags": ["users"]
      }
    },
    "/api/v1/users/me/contact/verify": {
      "post": {
        "operationId": "users_verifyContactChange",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyContactChangeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["users"]
      }
    },
    "/api/v1/users/me/summary": {
      "get": {
        "operationId": "users_summary",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "The \"who am I here\" block behind the avatar menu — see UsersService.summary.\n\nNot `@Public()`, though it was. `@CurrentUser()` reads `request.user`, which\nthe auth guard leaves unset on a public route, so an unauthenticated call\nreached Prisma with `userId: undefined` and answered 500. There is nothing\nfor a guest here in any case: every field describes an account.",
        "tags": ["users"]
      }
    },
    "/api/v1/users/{id}": {
      "get": {
        "operationId": "users_publicProfile",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["users"]
      }
    },
...
    "/api/v1/admin/users/{id}": {
      "get": {
        "operationId": "admin_userDetail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Read-only, any admin. Mutations below are super-admin only.",
        "tags": ["admin"]
      }
    },
...
  },
  "info": {
    "title": "FotSpot API",
...
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "The `accessToken` returned by any `/auth/*` login endpoint."
      }
    },
    "schemas": {
...
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "maxLength": 60
          },
          "lastName": {
            "type": "string",
            "maxLength": 60
          },
          "username": {
            "type": "string",
            "description": "The public handle. Accepted with or without a leading `@`; shape and\nreserved-word rules live in username.util.ts, uniqueness in the database.",
            "maxLength": 40
          },
          "avatarStorageKey": {
            "type": "string",
            "description": "Set only by confirming an upload from `POST /users/me/avatar/upload-url`.\nAccepting an arbitrary URL here would let a profile embed a remote image and\nturn every profile view into a request to a third-party server.",
            "maxLength": 500
          },
          "isPrivate": {
            "type": "boolean",
            "description": "Hide the account from public listings and profile reads. Off by default —\nan account nobody can find cannot be scouted."
          }
        }
      },
      "AvatarUploadUrlDto": {
        "type": "object",
        "properties": {
          "filename": {
            "type": "string",
            "maxLength": 200
          },
          "contentType": {
            "type": "string",
            "description": "Sent to R2 so the object is served back with the right type.",
            "maxLength": 100
          }
        },
        "required": ["filename"]
      },
      "VerificationChannel": {
        "type": "string",
        "enum": ["PHONE", "EMAIL"]
      },
      "RequestContactChangeDto": {
        "type": "object",
        "properties": {
          "channel": {
            "allOf": [
              {
                "$ref": "#/components/schemas/VerificationChannel"
              }
            ]
          },
          "destination": {
            "type": "string",
            "maxLength": 200
          }
        },
        "required": ["channel", "destination"]
      },
      "VerifyContactChangeDto": {
        "type": "object",
        "properties": {
          "channel": {
            "allOf": [
              {
                "$ref": "#/components/schemas/VerificationChannel"
              }
            ]
          },
          "destination": {
            "type": "string",
            "maxLength": 200
          },
          "code": {
            "type": "string",
            "minLength": 6,
            "maxLength": 6
          }
        },
        "required": ["channel", "destination", "code"]
      },
...
    }
  },
...
}